--- packages/genesys-spark-components/src/components/beta/gux-popover-beta/example.html
@@ -1,61 +1,0 @@
-<h1>gux-popover-beta</h1>
-<div class="scroll-container">
-  <div class="scroll-interior">
-    <div id="popover-target">Example Element</div>
-    <gux-popover-beta
-      id="popover-example"
-      position="top"
-      for="popover-target"
-      is-open
-    >
-      <span slot="title">Title</span>
-      <div style="width: 200px">
-        <div>popover content</div>
-        <gux-dropdown filterable placeholder="Select a country">
-          <gux-listbox aria-label="Countries">
-            <gux-option value="AFG">Afghanistan</gux-option>
-            <gux-option value="BHR">Bahrain</gux-option>
-            <gux-option value="ZWE">Zimbabwe</gux-option>
-          </gux-listbox>
-        </gux-dropdown>
-      </div>
-    </gux-popover-beta>
-  </div>
-</div>
-
-<style
-  onload="(function () {
-  const popoverTarget = document.getElementById('popover-target');
-  const popoverExample = document.getElementById('popover-example');
-
-  popoverTarget.addEventListener('mouseover', () => {
-    popoverExample.isOpen = true;
-  });
-})()"
->
-  .scroll-container {
-    width: 900px;
-    height: 300px;
-    margin: auto;
-    overflow: scroll;
-    border: 1px solid #aaa;
-  }
-
-  .scroll-interior {
-    position: relative;
-    width: 600px;
-    height: 600px;
-    background-color: #fff;
-  }
-
-  #popover-target {
-    position: absolute;
-    top: 250px;
-    left: 250px;
-    display: inline-block;
-    padding: 10px 15px;
-    background-color: #fdfdfd;
-    border: 1px #e4e9f0;
-    box-shadow: 0 0 2px fade(@gux-black-30, 24%);
-  }
-</style>
