--- packages/genesys-spark-components/src/components/beta/gux-popover-beta/gux-popover-beta.less
@@ -1,82 +1,0 @@
-@import (reference) '../../../style/color.less';
-@import (reference) '../../../style/shadows.less';
-@import (reference) '../../../style/spacing.less';
-@import (reference) '../../../style/zindex.less';
-@import (reference) '../../../style/typography.less';
-
-@gux-arrow-height: 5px;
-
-.gux-popover-wrapper {
-  position: absolute;
-  top: 0;
-  left: 0;
-  z-index: var(--gux-zindex-popover, @gux-zindex-popover);
-  display: inline-block;
-  padding: @gux-spacing-medium;
-  color: @gux-black-50;
-  background-color: @gux-grey-100;
-  border-radius: @gux-spacing-2xs;
-  .gux-shadow-40();
-
-  &.gux-hidden {
-    display: none;
-  }
-
-  .gux-arrow {
-    position: absolute;
-    width: @gux-arrow-height * 2;
-    height: @gux-arrow-height * 2;
-    background: inherit;
-  }
-
-  &[data-placement='bottom'],
-  &[data-placement='bottom-start'],
-  &[data-placement='bottom-end'] {
-    .gux-arrow {
-      border-top: 1px solid @gux-grey-30;
-      border-left: 1px solid @gux-grey-30;
-    }
-  }
-
-  &[data-placement='top'],
-  &[data-placement='top-start'],
-  &[data-placement='top-end'] {
-    .gux-arrow {
-      border-right: 1px solid @gux-grey-30;
-      border-bottom: 1px solid @gux-grey-30;
-    }
-  }
-
-  &[data-placement='left'],
-  &[data-placement='left-start'],
-  &[data-placement='left-end'] {
-    .gux-arrow {
-      border-top: 1px solid @gux-grey-30;
-      border-right: 1px solid @gux-grey-30;
-    }
-  }
-
-  &[data-placement='right'],
-  &[data-placement='right-start'],
-  &[data-placement='right-end'] {
-    .gux-arrow {
-      border-bottom: 1px solid @gux-grey-30;
-      border-left: 1px solid @gux-grey-30;
-    }
-  }
-
-  .gux-popover-header {
-    .body-font-bold();
-
-    display: flex;
-    flex-direction: row;
-    align-content: center;
-    align-items: center;
-    justify-content: space-between;
-    padding-bottom: @gux-spacing-medium;
-  }
-
-  gux-dismiss-button {
-    float: right;
-  }
-}
