+++ packages/genesys-spark-components/src/components/stable/gux-dismiss-button/readme.md
@@ -19,7 +19,6 @@ This component is meant for use in other components as a dismiss button.
  - [gux-modal](../gux-modal)
  - [gux-notification-toast-legacy](../../legacy/gux-notification-toast)
  - [gux-popover](../gux-popover)
- - [gux-popover-beta](../../beta/gux-popover-beta)
  - [gux-popover-list](../gux-popover-list)
  - [gux-popover-list-beta](../../beta/gux-popover-list-beta)
  - [gux-simple-toast-legacy](../../legacy/gux-simple-toast)
@@ -36,7 +35,6 @@ graph TD;
   gux-modal --> gux-dismiss-button
   gux-notification-toast-legacy --> gux-dismiss-button
   gux-popover --> gux-dismiss-button
-  gux-popover-beta --> gux-dismiss-button
   gux-popover-list --> gux-dismiss-button
   gux-popover-list-beta --> gux-dismiss-button
   gux-simple-toast-legacy --> gux-dismiss-button
