--- packages/genesys-spark-components/src/components/stable/gux-popover/gux-popover.types.ts
@@ -1,13 +1,0 @@
-export type PopperPosition =
-  | 'top'
-  | 'top-start'
-  | 'top-end'
-  | 'right'
-  | 'right-start'
-  | 'right-end'
-  | 'bottom'
-  | 'bottom-start'
-  | 'bottom-end'
-  | 'left'
-  | 'left-start'
-  | 'left-end';
