+++ packages/genesys-spark-components/src/components/stable/gux-popover/readme.md
@@ -12,12 +12,13 @@
 
 ## Properties
 
-| Property               | Attribute                | Description                                                                          | Type                                                                                                                                                                 | Default     |
-| ---------------------- | ------------------------ | ------------------------------------------------------------------------------------ | -------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ----------- |
-| `closeOnClickOutside`  | `close-on-click-outside` | Close popover when the user clicks outside of its bounds                             | `boolean`                                                                                                                                                            | `false`     |
-| `displayDismissButton` | `display-dismiss-button` | Indicate if the dismiss button is displayed                                          | `boolean`                                                                                                                                                            | `undefined` |
-| `for`                  | `for`                    | Indicates the id of the element the popover should anchor to                         | `string`                                                                                                                                                             | `undefined` |
-| `position`             | `position`               | Indicate position of popover element arrow (follow popper js position attribute api) | `"bottom" \| "bottom-end" \| "bottom-start" \| "left" \| "left-end" \| "left-start" \| "right" \| "right-end" \| "right-start" \| "top" \| "top-end" \| "top-start"` | `'bottom'`  |
+| Property               | Attribute                | Description                                                                             | Type                                                                                                                                                                 | Default     |
+| ---------------------- | ------------------------ | --------------------------------------------------------------------------------------- | -------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ----------- |
+| `closeOnClickOutside`  | `close-on-click-outside` | Close popover when the user clicks outside of its bounds                                | `boolean`                                                                                                                                                            | `false`     |
+| `displayDismissButton` | `display-dismiss-button` | Indicate if the dismiss button is displayed                                             | `boolean`                                                                                                                                                            | `undefined` |
+| `for`                  | `for`                    | Indicates the id of the element the popover should anchor to                            | `string`                                                                                                                                                             | `undefined` |
+| `isOpen`               | `is-open`                |                                                                                         | `boolean`                                                                                                                                                            | `false`     |
+| `position`             | `position`               | Indicate position of popover element arrow (follow floating ui placement attribute api) | `"bottom" \| "bottom-end" \| "bottom-start" \| "left" \| "left-end" \| "left-start" \| "right" \| "right-end" \| "right-start" \| "top" \| "top-end" \| "top-start"` | `'bottom'`  |
 
 
 ## Events
@@ -37,6 +38,10 @@
 
 ## Dependencies
 
+### Used by
+
+ - [gux-pagination-ellipsis-button](../../beta/gux-pagination-beta/gux-pagination-buttons-beta/gux-pagination-ellipsis-button)
+
 ### Depends on
 
 - [gux-dismiss-button](../gux-dismiss-button)
@@ -46,6 +51,7 @@
 graph TD;
   gux-popover --> gux-dismiss-button
   gux-dismiss-button --> gux-icon
+  gux-pagination-ellipsis-button --> gux-popover
   style gux-popover fill:#f9f,stroke:#333,stroke-width:4px
 ```
